requirements.txt
google-generativeai
tqdm 
cohere 
